PicoGraphics: Refactor pen types into their own cpp files.
