+++ libraries/pico_graphics/pico_graphics.cmake
@@ -1,6 +1,10 @@
 add_library(pico_graphics 
     ${CMAKE_CURRENT_LIST_DIR}/types.cpp
     ${CMAKE_CURRENT_LIST_DIR}/pico_graphics.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pico_graphics_pen_p4.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pico_graphics_pen_p8.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pico_graphics_pen_rgb332.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/pico_graphics_pen_rgb565.cpp
 )
 
 target_include_directories(pico_graphics INTERFACE ${CMAKE_CURRENT_LIST_DIR})
