+++ libraries/pico_graphics/pico_graphics.hpp
@@ -220,94 +220,15 @@ namespace pimoroni {
       uint8_t color;
       RGB palette[16];
 
-      PicoGraphics_PenP4(uint16_t width, uint16_t height, void *frame_buffer)
-      : PicoGraphics(width, height, frame_buffer) {
-        this->pen_type = PEN_P4;
-        if(this->frame_buffer == nullptr) {
-          this->frame_buffer = (void *)(new uint8_t[buffer_size(width, height)]);
-        }
-        for(auto i = 0u; i < 16; i++) {
-          palette[i] = {
-            uint8_t(i << 4),
-            uint8_t(i << 4),
-            uint8_t(i << 4)
-          };
-        }
-      }
-      void set_pen(uint c) {
-        color = c & 0xf;
-      }
-      void set_pen(uint8_t r, uint8_t g, uint8_t b) override {
-        int pen = RGB(r, g, b).closest(palette, 16);
-        if(pen != -1) color = pen;
-      }
-      int update_pen(uint8_t i, uint8_t r, uint8_t g, uint8_t b) override {
-        i &= 0xf;
-        palette[i] = {r, g, b};
-        return i;
-      }
-      void set_pixel(const Point &p) override {
-        // pointer to byte in framebuffer that contains this pixel
-        uint8_t *buf = (uint8_t *)frame_buffer;
-        uint8_t *f = &buf[(p.x / 2) + (p.y * bounds.w / 2)];
-
-        uint8_t  o = (~p.x & 0b1) * 4; // bit offset within byte
-        uint8_t  m = ~(0b1111 << o);   // bit mask for byte
-        uint8_t  b = color << o;       // bit value shifted to position
-
-        *f &= m; // clear bits
-        *f |= b; // set value
-      }
-      void set_pixel_dither(const Point &p, const RGB &c) override {
-        if(!bounds.contains(p)) return;
-        static uint pattern[16] = // dither pattern
-            {0, 8, 2, 10, 12, 4, 14, 6, 3, 11, 1, 9, 15, 7, 13, 5};
-
-        static std::array<uint8_t, 16> candidates;
-        get_dither_candidates(c, palette, 256, candidates);
-
-        // find the pattern coordinate offset
-        uint pattern_index = (p.x & 0b11) | ((p.y & 0b11) << 2);
-
-        // set the pixel
-        color = candidates[pattern[pattern_index]];
-        set_pixel(p);
-      }
-      void scanline_convert(PenType type, conversion_callback_func callback) override {
-        if(type == PEN_RGB565) {
-          // Cache the RGB888 palette as RGB565
-          RGB565 cache[16];
-          for(auto i = 0u; i < 16; i++) {
-            cache[i] = palette[i].to_rgb565();
-          }
-
-          // Treat our void* frame_buffer as uint8_t
-          uint8_t *src = (uint8_t *)frame_buffer;
-
-          // Allocate a per-row temporary buffer
-          uint16_t row_buf[bounds.w];
-          for(auto y = 0; y < bounds.h; y++) {
-            /*if(scanline_interrupt != nullptr) {
-              scanline_interrupt(y);
-              // Cache the RGB888 palette as RGB565
-              for(auto i = 0u; i < 16; i++) {
-                cache[i] = palette[i].to_rgb565();
-              }
-            }*/
-
-            for(auto x = 0; x < bounds.w; x++) {
-              uint8_t c = src[(bounds.w * y / 2) + (x / 2)];
-              uint8_t  o = (~x & 0b1) * 4; // bit offset within byte
-              uint8_t  b = (c >> o) & 0xf; // bit value shifted to position
-              row_buf[x] = cache[b];
-            }
-            // Callback to the driver with the row data
-            callback(row_buf, bounds.w * sizeof(RGB565));
-          }
-        }
-      }
+      PicoGraphics_PenP4(uint16_t width, uint16_t height, void *frame_buffer);
+      void set_pen(uint c) override;
+      void set_pen(uint8_t r, uint8_t g, uint8_t b) override;
+      int update_pen(uint8_t i, uint8_t r, uint8_t g, uint8_t b) override;
+      void set_pixel(const Point &p) override;
+      void set_pixel_dither(const Point &p, const RGB &c) override;
+      void scanline_convert(PenType type, conversion_callback_func callback) override;
       static size_t buffer_size(uint w, uint h) {
-        return w * h / 2;
+          return w * h / 2;
       }
   };
 
@@ -316,86 +237,14 @@ namespace pimoroni {
       uint8_t color;
       RGB palette[256];
       bool used[256];
-      PicoGraphics_PenP8(uint16_t width, uint16_t height, void *frame_buffer)
-      : PicoGraphics(width, height, frame_buffer) {
-        this->pen_type = PEN_P8;
-        if(this->frame_buffer == nullptr) {
-          this->frame_buffer = (void *)(new uint8_t[buffer_size(width, height)]);
-        }
-        for(auto i = 0u; i < 256; i++) {
-          palette[i] = {0, 0, 0};
-          used[i] = false;
-        }
-      }
-      void set_pen(uint c) override {
-        color = c;
-      }
-      void set_pen(uint8_t r, uint8_t g, uint8_t b) override {
-        int pen = RGB(r, g, b).closest(palette, 16);
-        if(pen != -1) color = pen;
-      }
-      int update_pen(uint8_t i, uint8_t r, uint8_t g, uint8_t b) override {
-        i &= 0xff;
-        palette[i] = {r, g, b};
-        return i;
-      }
-      int create_pen(uint8_t r, uint8_t g, uint8_t b) override {
-        // Create a colour and place it in the palette if there's space
-        for(auto i = 0u; i < 256u; i++) {
-          if(!used[i]) {
-            palette[i] = {r, g, b};
-            used[i] = true;
-            return i;
-          }
-        }
-        return -1;
-      }
-      int reset_pen(uint8_t i) {
-        palette[i] = {0, 0, 0};
-        used[i] = false;
-        return i;
-      }
-      void set_pixel(const Point &p) override {
-        uint8_t *buf = (uint8_t *)frame_buffer;
-        buf[p.y * bounds.w + p.x] = color;
-      }
-      void set_pixel_dither(const Point &p, const RGB &c) override {
-        if(!bounds.contains(p)) return;
-        static uint pattern[16] = // dither pattern
-            {0, 8, 2, 10, 12, 4, 14, 6, 3, 11, 1, 9, 15, 7, 13, 5};
-
-        static std::array<uint8_t, 16> candidates;
-        get_dither_candidates(c, palette, 256, candidates);
-
-        // find the pattern coordinate offset
-        uint pattern_index = (p.x & 0b11) | ((p.y & 0b11) << 2);
-
-        // set the pixel
-        color = candidates[pattern[pattern_index]];
-        set_pixel(p);
-      }
-      void scanline_convert(PenType type, conversion_callback_func callback) override {
-        if(type == PEN_RGB565) {
-          // Cache the RGB888 palette as RGB565
-          RGB565 cache[256];
-          for(auto i = 0u; i < 256; i++) {
-            cache[i] = palette[i].to_rgb565();
-          }
-
-          // Treat our void* frame_buffer as uint8_t
-          uint8_t *src = (uint8_t *)frame_buffer;
-
-          // Allocate a per-row temporary buffer
-          uint16_t row_buf[bounds.w];
-          for(auto y = 0; y < bounds.h; y++) {
-            for(auto x = 0; x < bounds.w; x++) {
-              row_buf[x] = cache[src[bounds.w * y + x]];
-            }
-            // Callback to the driver with the row data
-            callback(row_buf, bounds.w * sizeof(RGB565));
-          }
-        }
-      }
+      PicoGraphics_PenP8(uint16_t width, uint16_t height, void *frame_buffer);
+      void set_pen(uint c) override;
+      void set_pen(uint8_t r, uint8_t g, uint8_t b) override;
+      int update_pen(uint8_t i, uint8_t r, uint8_t g, uint8_t b) override;
+      int create_pen(uint8_t r, uint8_t g, uint8_t b) override;
+      int reset_pen(uint8_t i) override;
+      void set_pixel(const Point &p) override;
+      void scanline_convert(PenType type, conversion_callback_func callback) override;
       static size_t buffer_size(uint w, uint h) {
         return w * h;
       }
@@ -405,111 +254,14 @@ namespace pimoroni {
     public:
       RGB332 color;
       RGB palette[256];
-      PicoGraphics_PenRGB332(uint16_t width, uint16_t height, void *frame_buffer)
-      : PicoGraphics(width, height, frame_buffer) {
-        this->pen_type = PEN_RGB332;
-        if(this->frame_buffer == nullptr) {
-          this->frame_buffer = (void *)(new uint8_t[buffer_size(width, height)]);
-        }
-        for(auto i = 0u; i < 256; i++) {
-          palette[i] = RGB((RGB332)i);
-        }
-      }
-      void set_pen(uint c) override {
-        color = c;
-      }
-      void set_pen(uint8_t r, uint8_t g, uint8_t b) override {
-        color = rgb_to_rgb332(r, g, b);
-      }
-      int create_pen(uint8_t r, uint8_t g, uint8_t b) override {
-        return rgb_to_rgb332(r, g, b);
-      }
-      void set_pixel(const Point &p) override {
-        uint8_t *buf = (uint8_t *)frame_buffer;
-        buf[p.y * bounds.w + p.x] = color;
-      }
-      void set_pixel_dither(const Point &p, const RGB &c) override {
-        if(!bounds.contains(p)) return;
-        static uint8_t _odm[16] = {
-          0,  8,  2, 10,
-          12,  4, 14,  6,
-          3, 11,  1,  9,
-          15,  7, 13,  5
-        };
-
-        uint8_t _dmv = _odm[(p.x & 0b11) | ((p.y & 0b11) << 2)];
-
-        uint8_t red = c.r & 0b11000000;        // Two bits red
-        uint8_t red_r = c.r & 0b111111;        // Remaining six bits red
-        red_r >>= 2;                           // Discard down to four bit
-
-        uint8_t grn = (c.g & 0b11000000) >> 3; // Two bits green
-        uint8_t grn_r = c.g & 0b111111;        // Remaining six bits green
-        grn_r >>= 2;                           // Discard down to four bit
-
-        uint8_t blu = (c.b & 0b10000000) >> 6; // One bit blue
-        uint8_t blu_r = c.b & 0b1111111;       // Remaining seven bits green
-        blu_r >>= 3;                           // Discard down to four bit
-
-        color = red | grn | blu;
-        if(red_r > _dmv) color |= 0b00100000;
-        if(grn_r > _dmv) color |= 0b00000100;
-        if(blu_r > _dmv) color |= 0b00000001;
-
-        set_pixel(p);
-      }
-      void set_pixel_dither(const Point &p, const RGB565 &c) override {
-        if(!bounds.contains(p)) return;
-        RGB565 cs = __builtin_bswap16(c);
-        static uint8_t _odm[16] = {
-          0,  8,  2, 10,
-          12,  4, 14,  6,
-          3, 11,  1,  9,
-          15,  7, 13,  5
-        };
-
-        uint8_t _dmv = _odm[(p.x & 0b11) | ((p.y & 0b11) << 2)];
-
-        //                      RRRRRGGGGGGBBBBB
-        uint8_t red   = (cs & 0b1100000000000000) >> 8;  // Two bits grn
-        uint8_t red_r = (cs & 0b0011100000000000) >> 10; // Four bits cmp
-
-        uint8_t grn   = (cs & 0b0000011000000000) >> 6;  // Two bit grn
-        uint8_t grn_r = (cs & 0b0000000111100000) >> 5;  // Four bit cmp
-
-        uint8_t blu   = (cs & 0b0000000000010000) >> 3;  // Two bit blu
-        uint8_t blu_r = (cs & 0b0000000000001111);       // Four bit cmp
-
-        color = red | grn | blu;
-        //                          RRRGGGBB
-        if(red_r > _dmv) color |= 0b00100000;
-        if(grn_r > _dmv) color |= 0b00000100;
-        if(blu_r > _dmv) color |= 0b00000001;
-
-        set_pixel(p);
-      }
-      void scanline_convert(PenType type, conversion_callback_func callback) override {
-        if(type == PEN_RGB565) {
-          // Cache the RGB888 palette as RGB565
-          RGB565 cache[256];
-          for(auto i = 0u; i < 256; i++) {
-            cache[i] = palette[i].to_rgb565();
-          }
-
-          // Treat our void* frame_buffer as uint8_t
-          uint8_t *src = (uint8_t *)frame_buffer;
-
-          // Allocate a per-row temporary buffer
-          uint16_t row_buf[bounds.w];
-          for(auto y = 0; y < bounds.h; y++) {
-            for(auto x = 0; x < bounds.w; x++) {
-              row_buf[x] = cache[src[bounds.w * y + x]];
-            }
-            // Callback to the driver with the row data
-            callback(row_buf, bounds.w * sizeof(RGB565));
-          }
-        }
-      }
+      PicoGraphics_PenRGB332(uint16_t width, uint16_t height, void *frame_buffer);
+      void set_pen(uint c) override;
+      void set_pen(uint8_t r, uint8_t g, uint8_t b) override;
+      int create_pen(uint8_t r, uint8_t g, uint8_t b) override;
+      void set_pixel(const Point &p) override;
+      void set_pixel_dither(const Point &p, const RGB &c) override;
+      void set_pixel_dither(const Point &p, const RGB565 &c) override;
+      void scanline_convert(PenType type, conversion_callback_func callback) override;
       static size_t buffer_size(uint w, uint h) {
         return w * h;
       }
@@ -519,27 +271,11 @@ namespace pimoroni {
     public:
       RGB src_color;
       RGB565 color;
-      PicoGraphics_PenRGB565(uint16_t width, uint16_t height, void *frame_buffer)
-      : PicoGraphics(width, height, frame_buffer) {
-        this->pen_type = PEN_RGB565;
-        if(this->frame_buffer == nullptr) {
-          this->frame_buffer = (void *)(new uint8_t[buffer_size(width, height)]);
-        }
-      }
-      void set_pen(uint c) override {
-        color = c;
-      }
-      void set_pen(uint8_t r, uint8_t g, uint8_t b) override {
-        src_color = {r, g, b};
-        color = src_color.to_rgb565();
-      }
-      int create_pen(uint8_t r, uint8_t g, uint8_t b) override {
-        return rgb_to_rgb565(r, g, b);
-      }
-      void set_pixel(const Point &p) override {
-        uint16_t *buf = (uint16_t *)frame_buffer;
-        buf[p.y * bounds.w + p.x] = color;
-      }
+      PicoGraphics_PenRGB565(uint16_t width, uint16_t height, void *frame_buffer);
+      void set_pen(uint c) override;
+      void set_pen(uint8_t r, uint8_t g, uint8_t b) override;
+      int create_pen(uint8_t r, uint8_t g, uint8_t b) override;
+      void set_pixel(const Point &p) override;
       static size_t buffer_size(uint w, uint h) {
         return w * h * sizeof(RGB565);
       }
